// src/app/model/student.ts
export class Student {

    id: number;
    ime: string;
    prezime: string;
    adresa: string;
    telefon: string;
    datumRodjenja: string;
    pol: string;
    username: string;
    password: string;
    mestoId: number;
    indeks: string;
    radniStatus: string;
    zanimanje?: string;
    brojRacuna: string;
}